datasets/books/parser/requirements.txt
spacy>=3.7,<4.0
ebooklib>=0.18
beautifulsoup4>=4.12
lxml>=5.0
pyyaml>=6.0
